tests-uuid: Merge index.ts into abstract-uuid.ts
